requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2
django-bootstrap-v5==1.0.4
django-phonenumber-field==5.2.0
django-widget-tweaks==1.4.8
httpagentparser==1.9.1
humanize==3.7.1
idna==2.10
importlib-metadata==3.10.1
lxml==4.6.3
numpy==1.20.3
pdfkit==0.6.1
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2-binary==2.8.6
pycountry==20.7.3
pyotp==2.6.0
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.5
zipp==3.4.1
mutagen==1.45.1
boto3==1.9.136
